+++ cmd/dfuseeos/cli/trxdb-loader.go
@@ -25,9 +25,9 @@ func init() {
 			cmd.Flags().String("trxdb-loader-http-listen-addr", KvdbHTTPServingAddr, "Listen address for /healthz endpoint")
 			cmd.Flags().Int("trxdb-loader-parallel-file-download-count", 2, "Maximum number of files to download in parallel")
 			cmd.Flags().Bool("trxdb-loader-allow-live-on-empty-table", true, "[LIVE] force pipeline creation if live request and table is empty")
-			cmd.Flags().Bool("trxdb-loader-truncation-enabled", false, "Enables the creation of truncation marker on writes")
-			cmd.Flags().Uint64("trxdb-loader-truncation-each", 0, "Truncates data that is older the defined X block number")
-			cmd.Flags().Uint64("trxdb-loader-truncation-ttl", 0, "Truncates at every X block interval")
+			cmd.Flags().Bool("trxdb-loader-truncation-enabled", false, "Write truncation markers, and enable the automated purge of blocks past the window")
+			cmd.Flags().Uint64("trxdb-loader-truncation-purge-interval", 1000, "Interval of blocks between each purge.")
+			cmd.Flags().Uint64("trxdb-loader-truncation-window", 0, "When truncating, purge blocks older than this amount of blocks.")
 			return nil
 		},
 		FactoryFunc: func(runtime *launcher.Runtime) (launcher.App, error) {
@@ -47,8 +47,8 @@ func init() {
 				HTTPListenAddr:            viper.GetString("trxdb-loader-http-listen-addr"),
 				ParallelFileDownloadCount: viper.GetInt("trxdb-loader-parallel-file-download-count"),
 				EnableTruncationMarker:    viper.GetBool("trxdb-loader-truncation-enabled"),
-				TruncationTTL:             viper.GetUint64("trxdb-loader-truncation-ttl"),
-				PurgerInterval:            viper.GetUint64("trxdb-loader-truncation-each"),
+				TruncationWindow:          viper.GetUint64("trxdb-loader-truncation-window"),
+				PurgerInterval:            viper.GetUint64("trxdb-loader-truncation-purge-interval"),
 			}, &trxdbLoaderApp.Modules{
 				BlockFilter: runtime.BlockFilter.TransformInPlace,
 			}), nil